+++ ixmp/utils.py
@@ -1,10 +1,8 @@
-import collections
+from collections.abc import Iterable
 import logging
 
 import pandas as pd
 import six
-
-import ixmp
 
 
 # globally accessible logger
@@ -49,12 +47,12 @@ def isstr(x):
 
 def isscalar(x):
     """Returns True if x is a scalar"""
-    return not isinstance(x, collections.Iterable) or isstr(x)
+    return not isinstance(x, Iterable) or isstr(x)
 
 
 def islistable(x):
     """Returns True if x is a list but not a string"""
-    return isinstance(x, collections.Iterable) and not isstr(x)
+    return isinstance(x, Iterable) and not isstr(x)
 
 
 def check_year(y, s):
@@ -109,20 +107,12 @@ def filtered(df, filters):
     return df[mask]
 
 
-def import_timeseries(mp, data, model, scenario, version=None,
-                      firstyear=None, lastyear=None):
-    if not isinstance(mp, ixmp.Platform):
-        raise ValueError("{} is not a valid 'ixmp.Platform' instance".
-                         format(mp))
-
-    if version is not None:
-        version = int(version)
-    scen = ixmp.Scenario(mp, model, scenario, version)
-
-    df = ixmp.utils.pd_read(data)
+def import_timeseries(scenario, data, firstyear=None, lastyear=None):
+    """Import from a *data* file into *scenario*."""
+    df = pd_read(data)
     df = df.rename(columns={c: str(c).lower() for c in df.columns})
 
-    cols = ixmp.utils.numcols(df)
+    cols = numcols(df)
     years = [int(i) for i in cols]
     fyear = int(firstyear or min(years))
     lyear = int(lastyear or max(years))
@@ -130,12 +120,13 @@ def import_timeseries(mp, data, model, scenario, version=None,
     df = df[['region', 'variable', 'unit'] + cols]
     df.region = [x if x == 'World' else 'R11_' + x for x in df.region]
 
-    scen.check_out(timeseries_only=True)
-    scen.add_timeseries(df)
+    scenario.check_out(timeseries_only=True)
+    scenario.add_timeseries(df)
 
     annot = 'adding timeseries data from file {}'.format(data)
     if firstyear is not None:
         annot = '{} from {}'.format(annot, firstyear)
     if lastyear is not None:
         annot = '{} until {}'.format(annot, lastyear)
-    scen.commit(annot)
+
+    scenario.commit(annot)