+++ tests/test_integration.py
@@ -10,14 +10,14 @@ TS_DF_CLEARED = TS_DF.copy()
 TS_DF_CLEARED.loc[0, 2005] = np.nan
 
 
-def test_run_clone(tmpdir, test_data_path, caplog):
+def test_run_clone(test_mp, test_data_path, caplog):
     # this test is designed to cover the full functionality of the GAMS API
     # - initialize a new platform instance
     # - creates a new scenario and exports a gdx file
     # - runs the tutorial transport model
     # - reads back the solution from the output
     # - performs the test on the objective value and the timeseries data
-    mp = ixmp.Platform(tmpdir, dbtype='HSQLDB')
+    mp = test_mp
     scen = make_dantzig(mp, solve=test_data_path)
     assert np.isclose(scen.var('z')['lvl'], 153.675)
     pdt.assert_frame_equal(scen.timeseries(iamc=True), TS_DF)
@@ -41,14 +41,16 @@ def test_run_clone(tmpdir, test_data_path, caplog):
     # cloning with `keep_solution=False` and `first_model_year`
     # drops the solution and removes all timeseries not marked `meta=True`
     # in the model horizon (i.e, `year >= first_model_year`)
-    scen4 = scen.clone(keep_solution=False, first_model_year=2005)
+    with pytest.warns(UserWarning,
+                      match='first_model_year` is deprecated'):
+        scen4 = scen.clone(keep_solution=False, first_model_year=2005)
     assert np.isnan(scen4.var('z')['lvl'])
     pdt.assert_frame_equal(scen4.timeseries(iamc=True), TS_DF_CLEARED)
 
 
-def test_run_remove_solution(tmpdir, test_data_path):
+def test_run_remove_solution(test_mp, test_data_path):
     # create a new instance of the transport problem and solve it
-    mp = ixmp.Platform(tmpdir, dbtype='HSQLDB')
+    mp = test_mp
     scen = make_dantzig(mp, solve=test_data_path)
     assert np.isclose(scen.var('z')['lvl'], 153.675)
 
@@ -93,10 +95,10 @@ def get_distance(scen):
 
 def test_multi_db_run(tmpdir, test_data_path):
     # create a new instance of the transport problem and solve it
-    mp1 = ixmp.Platform(tmpdir / 'mp1', dbtype='HSQLDB')
+    mp1 = ixmp.Platform(backend='jdbc', driver='hsqldb', path=tmpdir / 'mp1')
     scen1 = make_dantzig(mp1, solve=test_data_path)
 
-    mp2 = ixmp.Platform(tmpdir / 'mp2', dbtype='HSQLDB')
+    mp2 = ixmp.Platform(backend='jdbc', driver='hsqldb', path=tmpdir / 'mp2')
     # add other unit to make sure that the mapping is correct during clone
     mp2.add_unit('wrong_unit')
     mp2.add_region('wrong_region', 'country')
@@ -113,7 +115,7 @@ def test_multi_db_run(tmpdir, test_data_path):
     del mp2
 
     # reopen the connection to the second platform and reload scenario
-    _mp2 = ixmp.Platform(tmpdir / 'mp2', dbtype='HSQLDB')
+    _mp2 = ixmp.Platform(backend='jdbc', driver='hsqldb', path=tmpdir / 'mp2')
     assert_multi_db(mp1, _mp2)
     scen2 = ixmp.Scenario(_mp2, **models['dantzig'])
 
@@ -131,10 +133,10 @@ def test_multi_db_run(tmpdir, test_data_path):
 
 def test_multi_db_edit_source(tmpdir):
     # create a new instance of the transport problem
-    mp1 = ixmp.Platform(tmpdir / 'mp1', dbtype='HSQLDB')
+    mp1 = ixmp.Platform(backend='jdbc', driver='hsqldb', path=tmpdir / 'mp1')
     scen1 = make_dantzig(mp1)
 
-    mp2 = ixmp.Platform(tmpdir / 'mp2', dbtype='HSQLDB')
+    mp2 = ixmp.Platform(backend='jdbc', driver='hsqldb', path=tmpdir / 'mp2')
     scen2 = scen1.clone(platform=mp2)
 
     pdt.assert_frame_equal(scen1.par('d'), scen2.par('d'))
@@ -156,10 +158,10 @@ def test_multi_db_edit_source(tmpdir):
 
 def test_multi_db_edit_target(tmpdir):
     # create a new instance of the transport problem
-    mp1 = ixmp.Platform(tmpdir / 'mp1', dbtype='HSQLDB')
+    mp1 = ixmp.Platform(backend='jdbc', driver='hsqldb', path=tmpdir / 'mp1')
     scen1 = make_dantzig(mp1)
 
-    mp2 = ixmp.Platform(tmpdir / 'mp2', dbtype='HSQLDB')
+    mp2 = ixmp.Platform(backend='jdbc', driver='hsqldb', path=tmpdir / 'mp2')
     scen2 = scen1.clone(platform=mp2)
 
     pdt.assert_frame_equal(scen1.par('d'), scen2.par('d'))