+++ RELEASE_NOTES.md
@@ -1,6 +1,23 @@
 
 # Next Release
 
+
+## Migration notes
+
+Configuration for ixmp and its storage backends has been streamlined.
+
+- Instead of ``DB_CONFIG_PATH``:
+  - Platform configuration is stored in the ``config.json`` configuration file.
+  - The :class:`Platform` constructor accepts the name of a stored platform configuration.
+  - Different storage backends may accept relative or absolute paths to backend-specific configuration files.
+- Instead of ``DEFAULT_DBPROPS_FILE``:
+  - On the command-line, use ``ixmp platform add NAME ARGS`` followed by ``ixmp platform add default NAME``.
+- Instead of ``DEFAULT_LOCAL_DB_PATH``: :obj:`ixmp.config` always contains a platform named 'local' that is located below the configuration path, in the directory 'localdb/default'.
+  - To change this path, use, e.g.: ``ixmp platform add local jdbc hsqldb PATH``.
+
+
+## All changes
+
 - [#182](https://github.com/iiasa/ixmp/pull/182): Add new Backend, Model APIs and JDBCBackend, GAMSModel classes.
 - [#188](https://github.com/iiasa/ixmp/pull/188),
   [#195](https://github.com/iiasa/ixmp/pull/195): Enhance reporting.